Moved code for title variants to lib
