+++ Importer/jctdata/jac.py
@@ -1,9 +1,9 @@
-import csv, json, itertools, string, os, re
-from unidecode import unidecode
+import csv, json, itertools, os, re
 from datetime import datetime
 
 from jctdata import settings
 from jctdata import resolver
+from lib.title_variants import title_variants
 
 ISSN_RX = "\d{4}-\d{3}[\dxX]"
 
@@ -31,6 +31,7 @@ def jac_index_data():
 
 
 def journals(coincident_issn_files, title_files, publisher_files):
+    print("JAC: Preparing journal autocomplete data")
     dir = datetime.strftime(datetime.utcnow(), settings.DIR_DATE_FORMAT)
     jacdir = os.path.join(settings.DATABASES, "jct", "jac", dir)
     os.makedirs(jacdir, exist_ok=True)
@@ -194,44 +195,15 @@ def _index(record):
     idx["issns"] = [issn.lower() for issn in record["issns"]]
     idx["issns"] += [issn.replace("-", "") for issn in idx["issns"]]
 
-    idx["title"] = _make_variants(record["title"])
+    idx["title"] = title_variants(record["title"])
 
     if "alts" in record:
         idx["alts"] = []
         for alt in record["alts"]:
-            idx["alts"] += _make_variants(alt)
+            idx["alts"] += title_variants(alt)
         idx["alts"] = list(set(idx["alts"]))
 
     record["index"] = idx
-
-
-def _make_variants(title):
-    title = title.lower()
-    variants = [title]
-    variants += _asciifold(title)
-    variants += [x for x in [_ampersander(t) for t in variants] if x is not None]
-    return list(set(variants))
-
-
-def _asciifold(val):
-    try:
-        asciititle = unidecode(val)
-    except:
-        asciititle = val
-
-    throwlist = string.punctuation + '\n\t'
-    unpunctitle = "".join(c for c in val if c not in throwlist).strip()
-    asciiunpunctitle = "".join(c for c in asciititle if c not in throwlist).strip()
-
-    return [unpunctitle, asciititle, asciiunpunctitle]
-
-
-def _ampersander(val):
-    if " & " in val:
-        return val.replace(" & ", " and ")
-    if " and " in val:
-        return val.replace(" and ", " & ")
-    return None
 
 
 if __name__ == "__main__":
